+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # LinDistFlow Changelog
 
+## develop
+- remove `dss_parse_lines` and `dss_parse_line_codes` methods; add `PowerModelsDistribution.parse_dss`
+- `Inputs` only requires a "master.dss" file path now (instead two file paths for line codes and lines)
+
 ## v0.1.4
 - update dependencies (JuMP 1.0, HiGHS for testing)